refactor linux containerenvvars code
